requirements.txt
Flask==2.0.1
h5py==3.1.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.3.2
protobuf==3.14.0
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.2
tensorflow-cpu==2.5.1
urllib3==1.26.5
Werkzeug==2.0.1
gunicorn==19.8.0
dnspython==2.1.0